requirements.txt
django
djangorestframework
django-cors-headers
PyJWT
mysqlclient
scikit-learn
fuzzywuzzy